Updates to screenshot code to not try uplaoding/converting if therimage isnt there and kill the one image after its created

+++ vps_kvm_screenshot.sh
@@ -12,14 +12,17 @@ else
   sleep 40 && kill $$
  }
  timer & timerpid=$!
- rm -f shot_$1.jpg shot1_$1.jpg;
+ rm -f shot1_$1.jpg;
  /root/cpaneldirect/vncsnapshot -dieblank -compresslevel 9 \
  -quality 100 -vncQuality 9 -allowblank -count 1 -fps 5 \
  -quiet 127.0.0.1:$display shot1_$1.jpg >/dev/null 2>&1;
- convert shot1_$1.jpg -quality 75 shot_$1.gif;
- rm -f shot1_$1.jpg;
- if [ ! "$url" = "" ] && [ -e "shot_$1.gif" ]; then
-  curl --connect-timeout 60 --max-time 240 -k -F screenshot=@shot_$1.gif "$url" 2>/dev/null;
+ if [ -e shot1_$1.jpg ]; then
+  convert shot1_$1.jpg -quality 75 shot_$1.gif;
+  rm -f shot1_$1.jpg;
+  if [ ! "$url" = "" ] && [ -e "shot_$1.gif" ]; then
+   curl --connect-timeout 60 --max-time 240 -k -F screenshot=@shot_$1.gif "$url" 2>/dev/null;
+  fi;
+  rm -f shot_$1.gif;
  fi
  kill "$timerpid"
 fi
